# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PGWOS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Habit {
  id: number;
  title: string;
  category: string;
  /** checkbox | timer | number | notes | text */
  inputType: string;
  isDefault: boolean;
  isActive: boolean;
  order: number;
  createdAt: string;
}

export interface CreateHabitBody {
  title: string;
  category: string;
  inputType: string;
  order?: number;
}

export interface UpdateHabitBody {
  title?: string;
  category?: string;
  inputType?: string;
  isActive?: boolean;
  order?: number;
}

export interface HabitLog {
  id: number;
  habitId: number;
  date: string;
  completed: boolean;
  /** @nullable */
  value: string | null;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  durationSeconds: number | null;
  createdAt: string;
}

export interface LogHabitBody {
  completed: boolean;
  /** @nullable */
  value?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
}

export interface HabitDaySummary {
  date: string;
  totalHabits: number;
  completedHabits: number;
  completionRate: number;
}

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  /** low | medium | high */
  priority: string;
  /** pending | in_progress | completed */
  status: string;
  /** @nullable */
  dueDate: string | null;
  /** @nullable */
  dueTime: string | null;
  isRecurring: boolean;
  createdAt: string;
  /** @nullable */
  completedAt: string | null;
}

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  priority: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  dueTime?: string | null;
  isRecurring?: boolean;
}

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  priority?: string;
  status?: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  dueTime?: string | null;
  isRecurring?: boolean;
}

export interface Lead {
  id: number;
  name: string;
  /** @nullable */
  businessName: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  city: string | null;
  /** @nullable */
  industry: string | null;
  /** @nullable */
  source: string | null;
  /** new | contacted | interested | follow_up | closed | lost */
  status: string;
  /** @nullable */
  nextFollowUpDate: string | null;
  /** @nullable */
  lastContactDate: string | null;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  dealValue: number | null;
  createdAt: string;
}

export interface CreateLeadBody {
  name: string;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  source?: string | null;
  status: string;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  dealValue?: number | null;
}

export interface UpdateLeadBody {
  name?: string;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  source?: string | null;
  status?: string;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  lastContactDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  dealValue?: number | null;
}

export interface CalendarEvent {
  id: number;
  title: string;
  /** task | meeting | follow_up | reminder */
  eventType: string;
  startDate: string;
  /** @nullable */
  startTime: string | null;
  /** @nullable */
  endDate: string | null;
  /** @nullable */
  endTime: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  color: string | null;
  createdAt: string;
}

export interface CreateEventBody {
  title: string;
  eventType: string;
  startDate: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface UpdateEventBody {
  title?: string;
  eventType?: string;
  startDate?: string;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface Note {
  id: number;
  title: string;
  content: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  title: string;
  content: string;
  tags?: string[];
}

export interface UpdateNoteBody {
  title?: string;
  content?: string;
  tags?: string[];
}

export interface Checkin {
  id: number;
  date: string;
  /** happy | neutral | low */
  mood: string;
  energyLevel: number;
  focusLevel: number;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CheckinOrNull {
  checkin: Checkin | null;
}

export interface CreateCheckinBody {
  mood: string;
  energyLevel: number;
  focusLevel: number;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardStats {
  streak: number;
  tasksCompletedToday: number;
  tasksTotalToday: number;
  habitCompletionRate: number;
  habitsCompletedToday: number;
  habitsTotalToday: number;
  totalLeads: number;
  activeLeads: number;
  closedLeads: number;
  lostLeads: number;
  productivityScore: number;
  messagesCount: number;
  followUpsToday: number;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  /** @nullable */
  lastCheckinDate: string | null;
}

export interface DailySummary {
  date: string;
  tasksCompleted: number;
  habitsCompleted: number;
  productivityScore: number;
}

export interface LeadStatusCount {
  status: string;
  count: number;
}

export interface DayStat {
  date: string;
  value: number;
}

export interface AnalyticsData {
  dailySummaries: DailySummary[];
  leadStatusCounts: LeadStatusCount[];
  taskCompletionTrend: DayStat[];
}

export interface Settings {
  id: number;
  userName: string;
  darkMode: boolean;
  morningReminderEnabled: boolean;
  afternoonReminderEnabled: boolean;
  eveningReminderEnabled: boolean;
  morningReminderTime: string;
  afternoonReminderTime: string;
  eveningReminderTime: string;
  createdAt: string;
}

export interface UpdateSettingsBody {
  userName?: string;
  darkMode?: boolean;
  morningReminderEnabled?: boolean;
  afternoonReminderEnabled?: boolean;
  eveningReminderEnabled?: boolean;
  morningReminderTime?: string;
  afternoonReminderTime?: string;
  eveningReminderTime?: string;
}

export type GetHabitHistoryParams = {
  days?: number;
};

export type ListTasksParams = {
  status?: string;
  priority?: string;
};

export type ListLeadsParams = {
  status?: string;
  search?: string;
};

export type ListEventsParams = {
  start?: string;
  end?: string;
};

export type ListNotesParams = {
  search?: string;
  tag?: string;
};

export type ListCheckinsParams = {
  limit?: number;
};

export type GetAnalyticsParams = {
  days?: number;
};