# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PGWOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  priority?: string;
  status?: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  dueTime?: string | null;
  isRecurring?: boolean;
}